v5.1
- Implemented video call support with image recognition using diffusion description
- Improved UI
- Updated UI to Kawai v5.1
- Fixed menus
- Improved mobile adaptivity
- Stabilized server works
- Full client solution
- Better dark mode
- Fixed overlaying CSS
- Project partial factoring
- Small directory renamings
- Uploaded datasets in the first beta
- Moved libraries in the "lib/" directory
- Fixed request bug
- Fixed overlaying topbar
- Improved voice recognition
