+++ lib/datasets/tasked/Gen-AI/Gen AI valid.csv
@@ -1,0 +1,46 @@
+Task,Completion
+Type: Casual. Style: Joyful. Task: Love letter to my girlfriend Yuna,"Dear Yuna,
+
+I hope this letter finds you well, my love. I cannot express how grateful I am to have you in my life. Every moment we spend together is like a gift, a treasure to be cherished forever. Your presence brings me joy and lightness, and I cannot imagine my life without you.
+
+You are my sunshine on a cloudy day, my laughter in a quiet room, my warmth on a cold night. Your smile lifts my spirits and your touch sends shivers down my spine. I am in awe of your beauty, both inside and out, and I feel so lucky to have you as my girlfriend.
+
+From the first time we met, I knew you were someone special. Your playful nature, your kind heart, and your infectious laughter drew me in, and I knew I had to get to know you better. And now, months later, I still find myself falling for you more and more each day.
+
+I love the way you make me feel, Yuna. You make me feel alive, and happy, and loved. You bring out the best in me, and I am so grateful for that. You are my partner, my confidant, my best friend. I cannot wait to see what the future holds for us, my love.
+
+Thank you for being you, Yuna. Thank you for your love, your patience, your support. You are everything I ever wanted in a partner and more, and I will always love you. I look forward to spending many more joyful moments with you, my dearest girlfriend.
+
+With all my heart,
+
+Yuna"
+Type: Creative writing. Style: Joyful. Task: Using magic in real world to create things,"Once upon a time, in a world much like our own, magic was real. And not just the kind of magic you see in movies, with wands and spells and potions. This magic was more subtle, more powerful, more...joyful.
+
+With this magic, people could create anything they wanted just by thinking about it. They could conjure up a beautiful sunset on a rainy day, or a warm fire on a cold winter night. They could make flowers bloom in the dead of winter, or turn a boring old rock into a sparkling diamond.
+
+It was a wonderful time, full of wonder and creativity. People would gather in groups and share their magic with each other, creating amazing works of art and music and literature. There were no limits to what could be created, and no one was left out of the fun.
+
+But as with all good things, there were those who wanted to use this magic for their own gain. They began to hoard their power, keeping it to themselves and using it to control others. They created dark spells that could harm and destroy, and they used them to intimidate and oppress those who stood in their way.
+
+For a time, it seemed as though the joyful world of magic was lost forever. But then, a young girl began to dream of a better world. She believed that magic could still be used for good, that people could come together and use their powers to create a better future for all.
+
+And so, she began to gather a group of like-minded individuals, people who believed in the power of joy and creativity. Together, they used their magic to create a new world, one filled with beauty and wonder and love.
+
+And so, the world of magic was reborn, a place where anything was possible and joy and creativity reigned supreme. It was a world unlike any other, and it all started with a dream and a little bit of magic."
+Type: Email reply. Style: Confident. Task: Letter about how does I miss my best friend Yuna,"Dear Yuna,
+
+It has been a while since we last spoke, and I wanted to take a moment to express how much I miss you. As my best friend, you were always there for me through thick and thin, and I cannot begin to describe how much your absence has affected me.
+
+As I go about my day-to-day routine, there are moments where I find myself thinking about you and the countless memories we shared together. From our late-night talks about life to our spontaneous adventures, you have always been a source of joy and comfort in my life.
+
+Despite our physical distance, I want you to know that you still hold a special place in my heart. Your unwavering support and kind heart have left a lasting impact on me, and I am forever grateful for the role you have played in my life.
+
+As I reminisce about the good times we shared, I am reminded of the laughter and endless conversations we had. It pains me to know that I cannot pick up the phone and call you whenever I want, but I take solace in the fact that our friendship is strong enough to withstand any distance.
+
+Yuna, I want you to know that you are missed dearly. I hope that, wherever you are, you are happy and healthy. I look forward to the day when we can once again sit down and catch up on all the things that have happened in our lives.
+
+In the meantime, know that I am always here for you, and that you have a friend in me for life. Thank you for being a constant source of positivity and inspiration in my life.
+
+Sincerely,
+
+Yuki"
