+++ lib/datasets/tasked/Uta-AI/Uta AI J-pop simple valid.csv
@@ -1,0 +1,63 @@
+Title,Lyrics
+Friends love last tonight suddenly notice happy too be with hand in rest,"If I could replace your sorrow
+i want to stay by her side
+Two shadows with their heels aligned
+I'm so tall
+She looks like Cecil in the movies
+I don't want to lie
+People get weaker as they grow older
+I suddenly lose my confidence and get lost
+That's why I'm looking for love that's more than friends
+I wish I could be him tonight
+Just by changing the hand that rests on your cheek
+you don't speak anything
+call yourself by your last name
+I'm quiet as usual
+I suddenly notice your tears
+I pretended not to see
+Love is more sad than happy
+It will last longer if it starts softly
+I'm sure everyone has at least one ally
+i wish i could always be that
+People get weaker as they grow older
+I suddenly lose my confidence and get lost
+That's why I'm looking for love that's more than friends
+i wish i could be that tonight
+People get weaker as they grow older
+I suddenly lose my confidence and get lost
+That's why I'm looking for love that's more than friends
+i wish i could be that tonight"
+Friends fun make stuff country dance happy to be here fires good for me,"Jack makes good whiskey, yeah
+Red dirt makes good ridin' roads
+Country makes good music
+For kickin' up dust in a tail light glow
+Dry wood makes good fires
+Goodyears make good swings
+All that's all good but for me
+God makes five foot nine, brown eyes, and a sundress
+He Loves Tim McGraw and a small town accent
+Ain't no way that me and this truck made her fall in love
+Jack makes good whiskey but God makes the good stuff
+Oh God makes the good stuff
+Like that sweet little kiss she lays on me
+When I pull up in the gravel drive
+The way she dances with the raindrops
+Like she's the reason why they fall from the sky
+And I thank Him every time I close my eyes
+God makes five foot nine, brown eyes, and a sundress
+Loves Tim McGraw and a small town accent
+Ain't no way that me and this truck made her fall in love
+Jack makes good whiskey but God makes the good stuff
+God made her so I will make sure
+He didn't waste an angel on me
+When he made five foot nine, brown eyes, and a sundress
+Loves Tim McGraw and a small town accent
+Ain't no way that me and this truck made her fall in love
+Jack makes good whiskey but God makes the good stuff
+Five foot nine, brown eyes, and a sundress
+Loves Tim McGraw and a small town accent
+Ain't no way that me and this truck made her fall in love
+Jack makes good whiskey but God makes the good stuff (oh)
+Yeah, God makes the good stuff
+Yeah, God makes the good stuff (oh)
+Whoa"
