+++ manifest.json
@@ -3,7 +3,7 @@
     "short_name": "Yuna",
     "description": "Yuna AI - your personal AI girlfriend",
     "lang": "en",
-    "start_url": "/",
+    "start_url": "",
     "display": "standalone",
     "background_color": "#99c189",
     "theme_color": "#99c189",
